Unlock width requests in board window GTK
The message field and clocks had a specified width to force them as
wide as the board irrespective of their text content. This prevented
the user to size the board window below the initial width. The size
requests are now removes after initial pop up, as they already were
for Graph widgets. This means the width of the window can now be
decreased, although we quickly run into the point where the texts
on clock or menu bar prevent further shrinking. (Which could only
be prevented by clipping the menu texts more, or decreasing the clock
font.)