Make user-adjusted board size quasi-persistent (GTK)
After adjusting the board-window size the newly calculated square size
will be used to derive a new -boardSize for saving in the settings file,
by rounding it down to the nearest standard size. (But the size actually
used for display will not be rounded!) This will cause the board to pop
up with approximately the same size next time. The remembered initial
square size will also be adapted to the rounded value, so that the fonts
will start to apply to the new -boardSize as well on saving settings.
This will only be done if the old -boardSize value did not contain
commas (meaning it specified more than just square size).