Adapt clock and message font after board-window sizing
The clock and message font are now adapted to the new -boardSize value
determined after sizing of the board window. When a valid font for the
new size was already available (e.g. read from the settings file), it
uses that. If not it uses the hard-coded default font name. In both
cases a %d in the name is expanded based on the size-table indication
for that font. This makes the clock font adpat even without prior history.