Fix explosion of clocks for large board size GTK
When packing things in a hbox for getting them on the same row,
the hbox was made vertically expandable when the height of the first
element was requested > 80 (assuming it would be a text memo). At
square sizes > 81 the logos would require a height > 80, though,
and would be the first element on the logo + clocks row. This would
make them expandable, so that the window height would be equally divided
between board and logos (which again would leave preciously little for
the board). Now the expandability is reserved for TextBox Options only.