Lock board size when clock changes to two lines
First writing in the clock widget will usually increase its height
because the clock font is bigger than default. And for two-line
clocks as used with logos this will even be more pronounced. This
did cause the extra size to be taken from the board, ending up with
a smaller square size than originally requested. Which again would
cause any font changes to be saved as belonging to that smaller board
size. So starting next time with the originally requested size (as
opposed to the saved board size) would then not use those fonts!
The board size is now locked during the first write to the clocks,
or writes that change the number of lines.