Conditionally replace 'other-window' fonts on sizing
Fonts not used in the main window (i.e. other than clock, coord and
message) are now also adjusted on sizing, but not when this would
mean replacing an explicitly specified font by a default one.
If we size to a boardSize for which a font was specified earlier,
we always switch to the latter, and if we were merely using the
default font for the old size we switch to the default font for
the new size rather than keeping it.