Store fonts changed by font dialog in fonts table
The fonts table holds fonts per type and size that should be saved
in the settings file because they were either read from it, or explicitly
specified by the user on the command line. This now also holds for
fonts altered through the Fonts dialog, without waiting for the saving
of settings, because by that time the boardSize might already be different.