Finish castling and e.p. rights for Edit Position
In stead of faking rights based on corner positioning of Rooks and
e-file King, EditPositionDone now interprets the rightsBoard that
recorded 'promotion clicks' on King, Rooks or Pawns to toggle their
rights. This rightsBoard is initialized from the castling rights in
the position that is being edited, and the clear and piece-pallette
boards now also set these castling rights correctly.