Mention support for Arena960 protocol with USI/UCCI checkbox WB
Normally the -uxiAdapter option would be configured to use UCI2WB,
and in the binary distribution of WinBoard it certainly will be.
As UCI2WB now supports the UCI dialect Arena uses to play Chess960,
people can use the USI/UCCI checkbox also for Chess960 engines that
use that protocol, and the checkbox label now points that out.