Extend full-board textures by periodic tiling (XB)
Texture images with an -NxM suffix to their name are scaled to match
that number of squares before cutting squares out of them. But if the
actual board was larger than NxM the remaining squares fell outside the
image and remained white. Now the cutting area is mapped back into
the image, effectively behaving as if it is periodically tiling the
entire plane.