Fix browsing for folders, and allow starting in DATADIR
The previous path seemed to cause a segfault when browsing for a folder
with an empty text-edit. In addition there existed a bug for browsing
from the Engine Settings dialog, where the value of the option was taken
as a filter. Both are fixed now.
 The Option.textValue field, which for FileName Options holds the filter,
for PatName Options could be NULL or "". The latter is now taken to
mean that browsing should start in DATADIR, and is only used when
browsing for sound files or pieceImageDir with empty pathname field.