Use missing SVG from parent if -pid name starts with sub_
If the current piece-image directory's name starts with "sub_", any
pieces not found in it will be searched in the parent directory.
This allows variants that need just a few pieces of a theme changed
to get their own sub_xxx sub-directory within the directory with SVG
images, containing the deviating SVGs. This will work recursively.