Limit prefilling with color to textures with alpha channel
When loading and preparing a texture we now test the format, and if
the image has an alpha channel we stay on the save side, and always
color the squares before drawing the texture. Only when the format
is not ARGB32 we skip the color fill. (And the default textures
do not have an alpha channel, so at least there no time will be
wasted on pointless color fills.)