Save programStartTime in settings file rather than save time
The -saveDate time stamp is used to decide whether XBoard has seen
options added to the master settings file before. But it looks only
to the master settings file at stratup. So when we save settings we
now put XBoards startup time as -saveDte in it, so that installs
when XBoard was already running would still find their way into the
user settings file.