Fix dragged piece during promotion popup
When dragging to a promotion square, the piece would keep 'hanging'
in the release position while the promotion popup was shown.
(This usually covered it, but not always.) As after a choice is made
the move will be animated, it was more logical to put it back on the
from-square.