Allow engine to specify holdings larger than board height
On reception of a 'setup' command we now automatically increase the
number of board ranks to the holdings size if the latter is larger,
but remember the number of added 'dead ranks'. When parsing a FEN these
ranks are blacked out, and on generating a FEN they are ignored. This
makes FENs behave like the board was not enlarged.